Bug-Fix : Fix get tile method for MVT Postgres for WorldCRS84Quad TMS
- `ST_TileEnvelope` PostGIS method did not work for WorldCRS84Quad TMS,
the MVTs generated with it were not valid tiles for the TMS. This is
probably because `ST_TileEnvelope` works properly with WebMercatorQuad,
and WebMercatorQuad and WorldCRS84Quad have different tile matrix sizes at the same zoom levels.

- Hence instead we manually find the coordinates of the tile envelope
based on the zoom level, x and y and use this in `ST_MakeEnvelope` to
create the tile envelope for the `ST_AsMVT` query.